+++ robotframework-ls/tests/robotframework_ls_tests/test_ast_utils.py
@@ -223,17 +223,12 @@ my
         for tok in node_info.node.tokens:
             if tok.type == tok.ARGUMENT:
                 found.append(
-                    f"is_kw {tok.value} = {handler.consider_current_argument_token_as_keyword(tok)}",
-                )
-                found.append(
-                    f"was_exp {tok.value} = {handler.was_last_expression_argument}",
+                    f"{tok.value} = {handler.next_tok_type_as_str(tok)}",
                 )
 
     assert found == [
-        "is_kw $v1 == ${v2} = False",
-        "was_exp $v1 == ${v2} = True",
-        "is_kw CallThis = True",
-        "was_exp CallThis = False",
+        "$v1 == ${v2} = <expression>",
+        "CallThis = <keyword>",
     ]
 
 
@@ -255,21 +250,69 @@ my
         for tok in node_info.node.tokens:
             if tok.type == tok.ARGUMENT:
                 found.append(
-                    f"is_kw {tok.value} = {handler.consider_current_argument_token_as_keyword(tok)}",
-                )
-                found.append(
-                    f"was_exp {tok.value} = {handler.was_last_expression_argument}",
+                    f"{tok.value} = {handler.next_tok_type_as_str(tok)}",
                 )
 
     assert found == [
-        "is_kw $v1 == ${v2} = False",
-        "was_exp $v1 == ${v2} = True",
-        "is_kw CallThis = True",
-        "was_exp CallThis = False",
-        "is_kw ELSE IF = False",
-        "was_exp ELSE IF = False",
-        "is_kw $v1 == ${v2} = False",
-        "was_exp $v1 == ${v2} = True",
-        "is_kw CallThis = True",
-        "was_exp CallThis = False",
+        "$v1 == ${v2} = <expression>",
+        "CallThis = <keyword>",
+        "ELSE IF = <control>",
+        "$v1 == ${v2} = <expression>",
+        "CallThis = <keyword>",
+    ]
+
+
+def test_run_keywords_1():
+    from robotframework_ls.impl.robot_workspace import RobotDocument
+    from robotframework_ls.impl import ast_utils
+
+    document = RobotDocument(
+        "uri",
+        """
+*** Keyword **
+my
+    Run Keywords   CallThis    Arg    AND    CallThis2""",
+    )
+    ast = document.get_ast()
+    found = []
+    for node_info in ast_utils.iter_indexed(ast, "KeywordCall"):
+        handler = ast_utils.get_args_as_keywords_handler(node_info.node)
+        for tok in node_info.node.tokens:
+            if tok.type == tok.ARGUMENT:
+                found.append(
+                    f"{tok.value} = {handler.next_tok_type_as_str(tok)}",
+                )
+
+    assert found == [
+        "CallThis = <keyword>",
+        "Arg = <none>",
+        "AND = <control>",
+        "CallThis2 = <keyword>",
+    ]
+
+
+def test_run_keywords_2():
+    from robotframework_ls.impl.robot_workspace import RobotDocument
+    from robotframework_ls.impl import ast_utils
+
+    document = RobotDocument(
+        "uri",
+        """
+*** Keyword **
+my
+    Run Keywords   CallThis    CallThis2""",
+    )
+    ast = document.get_ast()
+    found = []
+    for node_info in ast_utils.iter_indexed(ast, "KeywordCall"):
+        handler = ast_utils.get_args_as_keywords_handler(node_info.node)
+        for tok in node_info.node.tokens:
+            if tok.type == tok.ARGUMENT:
+                found.append(
+                    f"{tok.value} = {handler.next_tok_type_as_str(tok)}",
+                )
+
+    assert found == [
+        "CallThis = <keyword>",
+        "CallThis2 = <keyword>",
     ]